Now @with_checker works just like @with_checker()
That is to say, you don't need the parentheses
anymore if you are specifying the generator types
through default values to the arguments.

This was accomplished by replacing with_checker
with a Checker class, and making with_checker
a function that has a special case for when its
only argument is a function.

Also added a test for this feature.