Uncovered and fixed a MAJOR bug!!!
It turns out that none of the unit tests had been
making sure that the randomized tests were
actually being run!  A new suite of tests,
test_basics, has been added to make sure that
the test is run for the correct number of times.

Adding this suite uncovered a major bug where it
turned out that, in fact, no tests were being run!
The bug comes from the fact that izip stops when
one of its generators runs out, so when, say, no
keyword generators were specified this meant that
the iteration was never run even though argument
generators were specified;  replacing izip with
izip_longest fixes the issue.