Floats are now generated uniformly at all scales
Before, floats were generated uniformly within the
range -1e7 to +1e7.  This meant that 90% of the
time you'd get a value of order 1e6, and you would
only get a value of order, say, 1 less than 0.001%
of the time.  This is not the desired behavior
since in general one wants to test all scales.
Thus, the default behavior is now to generate
numbers whose *logarithm* is uniformly distributed
on, say, a scale from -7 to +7.  An additional
coin is then flipped to determine whether the
number should be positive or negative.

This change does not affect "frange" or
"unit_interval_float", which continue to provide
uniformly distributed floats.