Revert "Added support for Python 3."
This reverts commit add9993689ffd33bf87aebe84079f5d095a7baf9.

I was wrong when I thought that my changes worked just fine in Python 2.
It turns out that Exception.with_traceback only exists in Python 3, and
I missed this because I forgot for a moment that Python only checks these
things at run time so when the code was accepted by the Python 2
interpreter I mistakenly thought that this meant that the method did
exist in Python 2 after all.